+++ src/broadphase/broadphase_collision_manager.cpp
@@ -39,6 +39,35 @@
 
 namespace coal {
 
+namespace detail {
+struct CollisionCallBackFunctorWrapper : CollisionCallBackBase {
+  CollisionCallBackFunctorWrapper(const CollisionCallBackFunctor& functor)
+      : m_functor(&functor) {}
+
+  void init() override {}
+
+  bool collide(CollisionObject* o1, CollisionObject* o2) override {
+    return (*m_functor)(o1, o2);
+  }
+
+  CollisionCallBackFunctor const* m_functor;
+};
+
+struct DistanceCallBackFunctorWrapper : DistanceCallBackBase {
+  DistanceCallBackFunctorWrapper(const DistanceCallBackFunctor& functor)
+      : m_functor(&functor) {}
+
+  void init() override {}
+
+  bool distance(CollisionObject* o1, CollisionObject* o2,
+                Scalar& dist) override {
+    return (*m_functor)(o1, o2, dist);
+  }
+
+  DistanceCallBackFunctor const* m_functor;
+};
+}  // namespace detail
+
 //==============================================================================
 BroadPhaseCollisionManager::BroadPhaseCollisionManager()
     : enable_tested_set_(false) {
@@ -61,6 +90,50 @@ void BroadPhaseCollisionManager::update(CollisionObject* updated_obj) {
   COAL_UNUSED_VARIABLE(updated_obj);
 
   update();
+}
+
+//==============================================================================
+void BroadPhaseCollisionManager::collide(
+    CollisionObject* obj, const CollisionCallBackFunctor& fn) const {
+  detail::CollisionCallBackFunctorWrapper wrapper{fn};
+  this->collide(obj, &wrapper);
+}
+
+//==============================================================================
+void BroadPhaseCollisionManager::collide(
+    const CollisionCallBackFunctor& fn) const {
+  detail::CollisionCallBackFunctorWrapper wrapper{fn};
+  this->collide(&wrapper);
+}
+
+//==============================================================================
+void BroadPhaseCollisionManager::collide(
+    BroadPhaseCollisionManager* other_manager,
+    const CollisionCallBackFunctor& fn) const {
+  detail::CollisionCallBackFunctorWrapper wrapper{fn};
+  this->collide(other_manager, &wrapper);
+}
+
+//==============================================================================
+void BroadPhaseCollisionManager::distance(
+    CollisionObject* obj, const DistanceCallBackFunctor& fn) const {
+  detail::DistanceCallBackFunctorWrapper wrapper{fn};
+  this->distance(obj, &wrapper);
+}
+
+//==============================================================================
+void BroadPhaseCollisionManager::distance(
+    const DistanceCallBackFunctor& fn) const {
+  detail::DistanceCallBackFunctorWrapper wrapper{fn};
+  this->distance(&wrapper);
+}
+
+//==============================================================================
+void BroadPhaseCollisionManager::distance(
+    BroadPhaseCollisionManager* other_manager,
+    const DistanceCallBackFunctor& fn) const {
+  detail::DistanceCallBackFunctorWrapper wrapper{fn};
+  this->distance(other_manager, &wrapper);
 }
 
 //==============================================================================